requirements.txt
fastapi>=0.111.0
mangum>=0.17.0
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
boto3>=1.34.0
garminconnect>=0.2.0
uvicorn>=0.30.0
pydantic>=2.0.0
